Separate persistence cloud model into its own module.
