+++ legend-engine-xt-persistence-cloud-pure/src/main/resources/META-INF/services/org.finos.legend.pure.m3.serialization.filesystem.repository.CodeRepositoryProvider
@@ -1,0 +1,1 @@
+org.finos.legend.pure.code.core.CorePersistenceCloudCodeRepositoryProvider
