+++ legend-engine-xt-persistence-cloud-pure/src/main/resources/core_persistence_cloud.definition.json
@@ -1,0 +1,5 @@
+{
+  "name" : "core_persistence_cloud",
+  "pattern" : "(meta::external::persistence)(::.*)?",
+  "dependencies" : ["platform", "core", "core_external_shared", "core_persistence"]
+}
