+++ legend-engine-xt-persistence-cloud-pure/src/main/resources/core_persistence_cloud/external/persistence/aws/metamodel_diagram.pure
