+++ legend-engine-xt-persistence-cloud-pure/src/test/java/org/finos/legend/pure/code/persistence/cloud/Test_Pure_Persistence_Cloud.java
@@ -12,20 +12,20 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package org.finos.legend.pure.code.core.persistence;
+package org.finos.legend.pure.code.persistence.cloud;
 
 import junit.framework.TestSuite;
 import org.finos.legend.pure.code.core.compiled.test.PureTestBuilderHelper;
 import org.finos.legend.pure.m3.execution.test.TestCollection;
 import org.finos.legend.pure.runtime.java.compiled.execution.CompiledExecutionSupport;
 
-public class Test_Pure_Persistence
+public class Test_Pure_Persistence_Cloud
 {
     public static TestSuite suite()
     {
         CompiledExecutionSupport executionSupport = PureTestBuilderHelper.getClassLoaderExecutionSupport();
         TestSuite suite = new TestSuite();
-        suite.addTest(PureTestBuilderHelper.buildSuite(TestCollection.collectTests("meta::pure::persistence", executionSupport.getProcessorSupport(), ci -> PureTestBuilderHelper.satisfiesConditions(ci, executionSupport.getProcessorSupport())), executionSupport));
+        suite.addTest(PureTestBuilderHelper.buildSuite(TestCollection.collectTests("meta::external::persistence", executionSupport.getProcessorSupport(), ci -> PureTestBuilderHelper.satisfiesConditions(ci, executionSupport.getProcessorSupport())), executionSupport));
         return suite;
     }
 }
